omni-fit/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
stripe==11.2.0
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
httpx==0.27.2